Parachutes cluttering the ground since Arma Update 112555
Since the latest Arma 2 patch, (112555), the para_drop chutes are no
longer deleting when the AI land on our server. These changes set up an
array to collect every _chute created, then at the end of the script it
waits 45 seconds, (for the AI to safely land on the ground), then
deletes all of the _chutes.